refactor: move the remaining edition-agnostic request handling to the platform
Four things were in the edition only by accident of where they were first
written. None of them knows anything about the game.

**Impersonation.** `ImpersonationMiddleware` has no edition imports at all — the
`ImpersonationLog` model it writes moved to `gyrinx.site` in #2108, and the
docstrings in both `n23/core/middleware.py` and `gyrinx/middleware.py` claiming
otherwise were stale from that day. Middleware and the start/stop views move to
the platform. The URLs and their `core:` names stay exactly where they were, so
no link changes and no redirect is needed.

**`csrf_failure`.** Renders no template and touches no edition model. Its only
tie was `reverse("core:index")` for the no-referer fallback, which resolves to
`/` — the same thing `safe_redirect` already defaults to. A platform view has no
business reversing an edition URL name to find somewhere to send people, so the
explicit fallback is dropped rather than replaced.

**`site_banner`.** Already imported nothing but `gyrinx.site.models`.

**`trigger_discord_issue_action`.** Turns a message in the project's Discord into
a GitHub issue on this repository. That is project tooling, not game content, and
it now sits beside the API view that receives the Discord interaction.

Clears four `n23.` references from settings: `CSRF_FAILURE_VIEW`, the
`ImpersonationMiddleware` entry, and the `site_banner` context processor — plus
both references in `gyrinx/api/`. `notifications` is the one context processor
left, pending the `Notification` model itself.

⚠️ **Deployment:** moving the Discord task changes its `TaskRoute.path`, so its
Pub/Sub topic is renamed:

    prod--gyrinx.tasks--n23.core.tasks.trigger_discord_issue_action
 →  prod--gyrinx.tasks--gyrinx.api.tasks.trigger_discord_issue_action

The new topic is provisioned on Cloud Run startup; the old one is orphaned and
can be deleted once this is deployed. Watch the deploy for
`Failed to provision task infrastructure`.

Tests move with the code they cover — `test_csrf.py` and
`test_context_processors.py` are now platform tests. Suite is 4184 passed, 13
skipped, unchanged from before this stack.
