+++ gyrinx/context_processors.py
@@ -1,11 +1,18 @@
 """Platform context processors.
 
-``site_banner`` and ``notifications`` deliberately stay in
-``n23.core.context_processors``: they query the ``Banner`` and ``Notification``
-models, which are still edition models. They move here once those models do.
+``notifications`` is still in ``n23.core.context_processors``: it counts
+``Notification`` rows, and that model has not moved to the platform yet.
 """
 
+import logging
+
 from django.conf import settings
+from django.core.cache import cache
+from django.db import DatabaseError, InterfaceError, OperationalError
+
+from gyrinx.site.models import BANNER_CACHE_KEY, BANNER_CACHE_TIMEOUT, Banner
+
+logger = logging.getLogger(__name__)
 
 
 def gyrinx_debug(request):
@@ -16,7 +23,7 @@ def gyrinx_debug(request):
 def impersonation(request):
     """Expose impersonation state to templates.
 
-    Set by :class:`n23.core.middleware.ImpersonationMiddleware`. When
+    Set by :class:`gyrinx.middleware.ImpersonationMiddleware`. When
     ``is_impersonating`` is true, ``request.user`` is the impersonated user and
     ``impersonator`` is the real admin.
     """
@@ -24,3 +31,55 @@ def impersonation(request):
         "is_impersonating": getattr(request, "is_impersonating", False),
         "impersonator": getattr(request, "impersonator", None),
     }
+
+
+def site_banner(request):
+    """
+    Add the current live banner to the context, if any exists and hasn't been dismissed.
+
+    Note that this is disabled in tests by directly setting BANNER_CACHE_KEY to False
+    """
+    context = {"banner": None}
+
+    # Try to get banner from cache first
+    live_banner = cache.get(BANNER_CACHE_KEY)
+
+    if live_banner is None:
+        # Banner not in cache, fetch from database
+        try:
+            live_banner = Banner.objects.filter(is_live=True).first()
+            # Cache the result (even if None) to avoid repeated DB queries
+            cache.set(BANNER_CACHE_KEY, live_banner or False, BANNER_CACHE_TIMEOUT)
+        except Banner.DoesNotExist:
+            # This is expected when no banner exists
+            live_banner = None
+            cache.set(BANNER_CACHE_KEY, False, BANNER_CACHE_TIMEOUT)
+        except (DatabaseError, OperationalError, InterfaceError) as e:
+            # Database-related errors should be logged but not break the page
+            # Use warning level instead of exception to reduce noise in logs
+            logger.warning(
+                f"Database error while fetching site banner: {type(e).__name__}: {e}"
+            )
+            # Don't cache on database errors - try again next request
+            live_banner = None
+        except Exception:
+            # Log any unexpected errors but don't break page rendering
+            logger.exception("Unexpected error in site_banner context processor")
+            live_banner = None
+
+    # Check for False cached value (meaning no banner exists)
+    if live_banner is False:
+        live_banner = None
+
+    # Only show banner if user hasn't dismissed it
+    if live_banner:
+        # Check if session is available (may not be in ASGI error handling contexts)
+        if hasattr(request, "session"):
+            dismissed_banners = request.session.get("dismissed_banners", [])
+            if str(live_banner.id) not in dismissed_banners:
+                context["banner"] = live_banner
+        else:
+            # Show banner if we can't check dismissed status
+            context["banner"] = live_banner
+
+    return context
