+++ gyrinx/tests/test_context_processors.py
@@ -6,7 +6,7 @@ from django.core.cache import cache
 
 from gyrinx.site.models import Banner
 from gyrinx.site.models import BANNER_CACHE_KEY
-from n23.core.context_processors import site_banner
+from gyrinx.context_processors import site_banner
 
 
 @pytest.fixture
